backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
openai
python-dotenv
markdown-it-py
mdit-py-plugins
aiohttp
tiktoken
aiosqlite
